+++ Assets/_Scripts/ObjectModules/HealthAttributes.cs
@@ -1,0 +1,30 @@
+﻿using UnityEngine;
+
+public enum ElementType { Physical, Fire, Ice, Poison }
+
+[System.Serializable]
+public class HealthAttributes {
+
+    [SerializeField] private AttributeCurves curves;
+
+    public int health;
+    [Range(0, 1)] public float defense;
+    public float healModifier;
+
+    public HealthAttributes(AttributeCurves curves) {
+        this.curves = curves;
+    }
+
+    public int ComputeDamage(int amount) {
+        return Mathf.FloorToInt(amount * curves.defenseCurve.Evaluate(defense));
+    }
+
+    public int ComputeHeal(int amount) {
+        return Mathf.FloorToInt(amount * curves.healModCurve.Evaluate(healModifier));
+    }
+}
+
+public class AttributeCurves {
+    public AnimationCurve defenseCurve, healModCurve;
+    public AttributeCurves Clone() => (AttributeCurves) MemberwiseClone();
+}
