+++ Assets/_Scripts/ObjectModules/HealthMeter.cs
@@ -2,7 +2,7 @@
 
 public abstract class HealthMeter : MonoBehaviour {
 
-    public void TakeDamage(int amount, ElementType element = ElementType.Physical) {
+    public void TakeDamage(int amount) {
 
     }
 }
