+++ Assets/_Scripts/Settings/DefaultAttributeCurves.cs
@@ -1,0 +1,8 @@
+﻿#if UNITY_EDITOR
+using UnityEngine;
+
+public class DefaultAttributeCurves : ScriptableObject {
+    [SerializeField] private AttributeCurves data;
+    public AttributeCurves Data => data.Clone();
+}
+#endif
