+++ Assets/_Scripts/Settings/DefaultAttributeCurves.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b4982aeb7bd7d2740ba760fd23646d35
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
