+++ Assets/_Scripts/Settings/DefaultSummonProperties.cs
@@ -1,7 +1,6 @@
 ﻿#if UNITY_EDITOR
 using UnityEngine;
 
-[CreateAssetMenu()]
 public class DefaultSummonProperties : ScriptableObject {
     public Material fadeMaterial;
     public AnimationCurve growthCurveXZ, growthCurveY;
