+++ Assets/_Scripts/Settings/Editor/DefaultAttributeCurves.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 1f8275217f97f7c438b972abc1fb040f
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
